requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cycler==0.10.0
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.7
djangorestframework==3.12.4
filelock==3.3.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.1
gunicorn==20.1.0
heroku==0.1.4
httplib2==0.19.1
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mcpi==1.2.0
mpmath==1.2.1
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
peewee==3.14.4
Pillow==8.3.1
platformdirs==2.4.0
pluggy==1.0.0
protobuf==3.18.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.1
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==1.5
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.25
sqlparse==0.4.2
sympy==1.8
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.7
virtualenv==20.8.1
virtualenvwrapper-win==1.2.6
Werkzeug==2.0.1
whitenoise==5.3.0
WTForms==2.3.3
psycopg2==2.7.3
